+++ packages/analysis/rollup.config.js
@@ -44,8 +44,28 @@ function glsl() {
     };
 }
 
+function wgsl() {
+    return {
+        transform(code, id) {
+            if (/\.wgsl$/.test(id) === false) return null;
+            let transformedCode = JSON.stringify(code.trim()
+                // .replace(/(^\s*)|(\s*$)/gm, '')
+                .replace(/\r/g, '')
+                .replace(/[ \t]*\/\/.*\n/g, '') // remove //
+                .replace(/[ \t]*\/\*[\s\S]*?\*\//g, '') // remove /* */
+                .replace(/\n{2,}/g, '\n')); // # \n+ to \n;;
+            transformedCode = `export default ${transformedCode};`;
+            return {
+                code: transformedCode,
+                map: { mappings: '' }
+            };
+        }
+    };
+}
+
 const basePlugins = [
     glsl(),
+    wgsl(),
     resolve({
         module : true,
         jsnext : true,
@@ -83,12 +103,7 @@ module.exports = [
         plugins: basePlugins.concat(production ? [
             terser({
                 output : { comments : '/^!/', beautify: true },
-                mangle : {
-                    properties: {
-                        'regex' : /^_/,
-                        'keep_quoted' : true
-                    }
-                }
+                mangle : true
             })
         ] : []),
         external : ['maptalks', '@maptalks/gltf-loader', '@maptalks/gl', '@maptalks/vt'],