+++ packages/layer-video/rollup.config.js
@@ -10,13 +10,7 @@ const plugins = [].concat(
     production
         ? [
               terser({
-                  mangle: {
-                      properties: {
-                          regex: /^_/,
-                          keep_quoted: true,
-                          reserved: ["on", "once", "off"],
-                      },
-                  },
+                  mangle: true,
                   output: {
                       keep_quoted_props: true,
                       beautify: false,
@@ -38,29 +32,43 @@ outro = `typeof console !== 'undefined' && console.log('${outro}');`;
 function glsl() {
     return {
         transform(code, id) {
-            if (
-                /\.vert$/.test(id) === false &&
-                /\.frag$/.test(id) === false &&
-                /\.glsl$/.test(id) === false
-            )
-                return null;
-            let transformedCode = code
-                .replace(/[ \t]*\/\/.*\n/g, "") // remove //
-                .replace(/[ \t]*\/\*[\s\S]*?\*\//g, "") // remove /* */
-                .replace(/\n{1,}/g, "\\n") // # \n+ to \n
-                .replace(/\r{1,}/g, "\\n") // # \r+ to \n
-                .replace(/"/g, '\\"');
-            transformedCode = `export default "${transformedCode}";`;
+            if (/\.vert$/.test(id) === false && /\.frag$/.test(id) === false && /\.glsl$/.test(id) === false) return null;
+            let transformedCode = JSON.stringify(code.trim()
+                .replace(/\r/g, '')
+                .replace(/[ \t]*\/\/.*\n/g, '') // remove //
+                .replace(/[ \t]*\/\*[\s\S]*?\*\//g, '') // remove /* */
+                .replace(/\n{2,}/g, '\n')); // # \n+ to \n;;
+            transformedCode = `export default ${transformedCode};`;
             return {
                 code: transformedCode,
-                map: { mappings: "" },
+                map: { mappings: '' }
             };
-        },
+        }
+    };
+}
+
+function wgsl() {
+    return {
+        transform(code, id) {
+            if (/\.wgsl$/.test(id) === false) return null;
+            let transformedCode = JSON.stringify(code.trim()
+                // .replace(/(^\s*)|(\s*$)/gm, '')
+                .replace(/\r/g, '')
+                .replace(/[ \t]*\/\/.*\n/g, '') // remove //
+                .replace(/[ \t]*\/\*[\s\S]*?\*\//g, '') // remove /* */
+                .replace(/\n{2,}/g, '\n')); // # \n+ to \n;;
+            transformedCode = `export default ${transformedCode};`;
+            return {
+                code: transformedCode,
+                map: { mappings: '' }
+            };
+        }
     };
 }
 
 const basePlugins = [
     glsl(),
+    wgsl(),
     resolve({
         module: true,
         jsnext: true,
@@ -96,12 +104,7 @@ module.exports = [
                 ? [
                       terser({
                           output: { comments: "/^!/" },
-                          mangle: {
-                              properties: {
-                                  regex: /^_/,
-                                  keep_quoted: true,
-                              },
-                          },
+                          mangle: true
                       }),
                   ]
                 : []