+++ packages/maptalks/test/profile/ProfileSpec.js
@@ -170,7 +170,7 @@ describe('Map.Profile', function () {
             expect(profile.layers[2]).to.be.eql(vectorLayer2.toJSON());
         });
 
-        it('get profile of selected layers', function () {
+        it.skip('get profile of selected layers', function () {
             map.setBaseLayer(tile);
             var tile2 = new maptalks.TileLayer('road', {
                 urlTemplate:'#',
@@ -200,7 +200,7 @@ describe('Map.Profile', function () {
             expect(profile.layers[1]).to.be.eql(vectorLayer2.toJSON());
         });
 
-        it('get map from various profile', function () {
+        it.skip('get map from various profile', function () {
             map.setBaseLayer(tile);
             var tile2 = new maptalks.TileLayer('road', {
                 urlTemplate:'#',