+++ packages/traffic/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @maptalks/traffic
 
+## 0.120.0
+
+### Minor Changes
+
+- first version of webgpu for integration test
+
+### Patch Changes
+
+- Updated dependencies
+  - @maptalks/gltf-layer@0.120.0
+
 ## 0.119.0
 
 ### Minor Changes