+++ packages/vt/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @maptalks/vt
 
+## 0.120.0
+
+### Minor Changes
+
+- first version of webgpu for integration test
+
+### Patch Changes
+
+- Updated dependencies
+  - @maptalks/gl@0.120.0
+  - @maptalks/gpu@0.120.0
+  - @maptalks/vt-plugin@0.120.0
+
 ## 0.119.0
 
 ### Minor Changes